move IL Voodoo and QueryMetadata out of the way
