+++ Machete.Data/Helpers/SqlServerUtils.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using Machete.Data.Dynamic;
 using Microsoft.EntityFrameworkCore;
 
 namespace Machete.Data.Helpers
