+++ Machete.Data/MacheteContext.cs
@@ -30,6 +30,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Diagnostics;
 using System.Linq;
 using System.Text;
+using Machete.Data.Dynamic;
 using Machete.Data.Helpers;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
